backend/requirements.txt
Django~=4.2.0
psycopg2-binary~=2.9.0
djangorestframework~=3.14.0
python-dotenv~=0.21.0
djangorestframework-simplejwt~=5.3.1
django-cors-headers~=4.3.1
pytesseract~=0.3.13